services/mcp-lambda-ecr/lambdas/mcp-client/requirements.txt
google-genai
requests
python-dotenv # ローカル開発用、本番Lambda環境では環境変数で設定
